add casts from size_t to int for printf
git-svn-id: http://arthurdejong.org/svn/nss-pam-ldapd/nss-pam-ldapd@1534 ef36b2f9-881f-0410-afb5-c4e39611909c

+++ tests/test_tio.c
@@ -174,8 +174,8 @@ static int test_blocks(size_t wbs, size_t wbl, size_t rbs, size_t rbl)
   /* set up the socket pair */
   assertok(socketpair(AF_UNIX,SOCK_STREAM,0,sp)==0);
   /* log */
-  printf("test_tio: writing %d blocks of %d bytes (%d total)\n",wbl,wbs,wbl*wbs);
-  printf("test_tio: reading %d blocks of %d bytes (%d total)\n",rbl,rbs,rbl*rbs);
+  printf("test_tio: writing %d blocks of %d bytes (%d total)\n",(int)wbl,(int)wbs,(int)(wbl*wbs));
+  printf("test_tio: reading %d blocks of %d bytes (%d total)\n",(int)rbl,(int)rbs,(int)(rbl*rbs));
   /* start the writer thread */
   wargs.fd=sp[0];
   wargs.blocksize=wbs;
